backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
Pillow>=10.0
python-decouple>=3.8
openai>=1.0.0
ollama>=0.1.7
